F1: corrige busca da home, quinteto na régua, skip link e higiene de dados
Análise sênior da F1 (motor determinístico + portal). Backend verificado
saudável (cadeia de auditoria íntegra, RN-001..005 corretas, contrato
frontend↔backend consistente). Corrigido o que não funcionava:

- web/app/consulta: ?q= da home agora resolve o município pela busca
  tolerante a acento do servidor (RF-PORTAL-001, antes sem consumidor no
  frontend) — a caixa de busca da home era um recurso morto
- web/components/ReguaProcedencia: acrescenta licença e hash à régua —
  o quinteto de procedência agora é indissociável do valor (RF-PORTAL-004)
- web/app/layout + globals.css: skip link ".skip-link" desce ao receber
  foco (-48px → 8px), antes ficava invisível em left:-9999 (WCAG 2.4.7)
- api/indicadores.cobertura + web/cobertura: count(o.*)::int, tipo number

Verificado: suíte 44/44 em itmt_teste recriado; portal com dados reais —
q=cuiaba → Cuiabá pré-selecionado; régua com fonte·ref·extração·lic·hash;
Tab revela o skip link; cadeia íntegra (345 eventos). Também removida do
banco dev a poluição de 4 "Indicador de teste" (runtime local, não o repo).

+++ api/src/indicadores/indicadores.service.ts
@@ -306,7 +306,7 @@ export class IndicadoresService {
       `SELECT m."Municipio_CodigoIbge" AS codigo_ibge, m."Municipio_Nome" AS municipio,
               t."TemaConsulta_Id" AS tema_id, t."TemaConsulta_Nome" AS tema,
               max(o."Observacao_DataReferencia")::text AS ultima_referencia,
-              count(o.*) AS observacoes
+              count(o.*)::int AS observacoes
          FROM "Municipio" m
         CROSS JOIN "TemaConsulta" t
          LEFT JOIN "SubtemaConsulta" s ON s."SubtemaConsulta_TemaId" = t."TemaConsulta_Id"
